Fix large_size_array test

+++ tests/behaviour/contracts/largeSizeArray/large_size_array.sol
@@ -1,13 +1,13 @@
 pragma solidity ^0.8.0;
 
 contract WARP {
-    uint256[100000000] public data;
+    uint8[100000000] public data;
 
     function getLength() public returns (uint256) {
         return data.length;
     }
 
-    function setValue(uint256 index, uint256 value) public returns (uint256) {
+    function setValue(uint8 index, uint8 value) public returns (uint8) {
         data[index] = value;
         return data[index];
     }
